pyproject.toml
[tool.poetry]
name = "talk-to-manifesto"
version = "0.1.0"
description = ""
authors = ["Sourav Kumar Bose <sourav.kumar.bose@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
scipy = "^1.14.1"
python-dotenv = "^1.0.1"
huggingface-hub = {extras = ["cli"], version = "^0.25.0"}
llvmlite = "^0.43.0"
librosa = "^0.10.2.post1"
pre-commit = "^3.8.0"
isort = "^5.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
